fix: Fixed IndexOutOfRange exception when building a project with analyzers
When building a project with analyzers, Roslyn expects that SyntaxTree length equals to the sourceFileAnalyzerConfigOptions length. However, because we add one more script to SyntaxTree, their lengths don't match and it causes the IndexOutOfRange exception. We don't need to check autogenerated files with analyzers, so skipping them resolves the issue.